Move all application code to an application subpackage
