+++ main.py
@@ -1,1 +1,1 @@
-from autogpt import main
+
