+++ tests/unit/test_config.py
@@ -8,8 +8,8 @@ from unittest.mock import patch
 
 import pytest
 
+from autogpt.app.configurator import GPT_3_MODEL, GPT_4_MODEL, create_config
 from autogpt.config import Config, ConfigBuilder
-from autogpt.configurator import GPT_3_MODEL, GPT_4_MODEL, create_config
 from autogpt.workspace.workspace import Workspace
